// frontend/src/app/learncoreservices/models/role-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface RoleResponse {
  createdDate?: string;
  description?: string;
  lastModifiedDate?: string;
  level?: 'PLATFORM' | 'TENANT';
  name?: string;
  permissionCodes?: Array<string>;
  tenantUuid?: string;
  uuid?: string;
}